refactory : homescreen 모듈화 (애니메이션 및 훅 분리)
